Fix H1 for SEO
Every page should have only one H1 tag. To achieve this, all H tags have been downgradded by one level, plus the content of map tooltips wrapped in an article tag (to not break the heading level structure).

As there is no real place to put the H1 tag (holding the text "Radlkarte [Active Region]") it is placed in the legend and colored the same as the legend tab header background color. This way it is visually hidden in a way Google should not detect, because Google doesn't like visually hidden H1s.